Fix idRef completions (#208)
The original implementation didn't check members for the idRef trait,
and in general didn't handle idRefs in trait values. This commit fixes
both of those issues by adding a `targetOf` field to TerminalShape, so
that member can be checked for the idRef trait, and uses ShapeCompleter
when a trait value completion is for an idRef.

I also added some extra tests for definition and hover to make sure it
was working correctly there too.