Fix mixin parsing (#211)
There could have been an infinite loop if you had something that wasn't
an identifier in a mixin, like:
```
structure Foo with [123] {}
structure Foo with 123 {}
```
so I changed the parser to recover to either an identifier, or some
structural breakpoint when that happens. I don't think this specific
strategy generalizes enough to just put it in `ident()`, but if we find
any other cases where this can happen we could add some more error
recover methods, or specific `ident()` methods that recover in a
particular way.

I also modified the invalid test cases in IdlParserTest to include
blocks - for some reason I had the test filtering out blocks before the
assertion, but we definitely want to check for them.

I also removed an erroneous `ShapeNode` Statement type.