Simplify NodeSearch (#215)
Previously, NodeSearch used NodeCursor::setCheckpoint to store the
current position in the cursor, inspect previous edges for computing
dynamic member targets, and return to the checkpoint to continue
traversal. Maybe I thought there were complex use-cases that had to look
further back in the path, but right now all we need is to look at the
previous Node. So I updated NodeSearch to use the previous Node to
compute dynamic member targets, and removed the checkpoint stuff from
NodeCursor.