Add changelog to github release notes
So we don't have to manually add the release notes to the github release
after it is published.

I added a custom task to read in the full changelog, and parse out the
entry for the latest version. The result is written out to a resources
file that I pointed the github release config to.

The jreleaser docs don't seem to be 100% clear on whether the changelog
is actually what ends up in the release notes, but it would make sense
to me, and the text in https://jreleaser.org/guide/latest/reference/release/github.html#_release_notes
suggests it.