Add runtime plugin for generating stand-alone images (#159)
* add runtime plugin for generating stand-alone images

* add JReleaser configuration for github release and runtimes

* Use runtime plugin fork

Updates the runtime plugin to a fork with support for java 21, instead
of the upstream https://github.com/beryx/badass-runtime-plugin. See
https://github.com/beryx/badass-runtime-plugin/pull/154 for more
details.

Also changed the java version back to 21.

* Remove deploy github workflow

We had a github workflow that would run on a tag push, creating a new
release and uploading a zip of the language server. Jreleaser should
take care of all that now, so the workflow is unnecessary.

* Add changelog to github release notes

So we don't have to manually add the release notes to the github release
after it is published.

I added a custom task to read in the full changelog, and parse out the
entry for the latest version. The result is written out to a resources
file that I pointed the github release config to.

The jreleaser docs don't seem to be 100% clear on whether the changelog
is actually what ends up in the release notes, but it would make sense
to me, and the text in https://jreleaser.org/guide/latest/reference/release/github.html#_release_notes
suggests it.

---------

Co-authored-by: Miles Ziemer <mziemer@amazon.com>