Add argument parser feature to the LSP. (#218)
Add the argument parser feature to the LSP and modified the Main class.

The argument parser will be able to parse the help and port-number arguments for now. For port-number argument, both flag argument and positional argument are supported. Port number is optional, and defaults to `0` to use standard in/out. The positional port argument is deprecated, but still supported.