Cleanup build script (#222)
Our build script was a mess and it has been bugging me for a while, so I'm
finally cleaning it up. I mostly reorganized things, and removed unnecessary
configuration, but there are some minor changes to note:
- Got rid of the git-version plugin, which shouldn't be necessary since we
  use the VERSION file to manage the project version.
- Got rid of the manual checkstyle dependency, since what we were setting
   it to was already the default.
- Changed test logging to only log failed and skipped, instead of passed and
   stdout/err, which should reduce some noise.
- Fixed the dependency declaration for smithy libraries, which weren't using
   string interpolation, so were literally "[smithyVersion, 2.0[". Either that was
   a thing you could do in groovy, or it's been busted since 2023. Yikes.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.